delete dml.rs
